+++ src/nm-dbus-manager.h
